image_process/requirements.txt
aiohttp
requests
numpy==1.26.4
Pillow==11.1.0
mediapipe==0.10.21